+++ Changelog.md
@@ -1,6 +1,7 @@
 ## development
 
   - Add your change HERE
+  - Adds FFaker::Date.birthday [@professor]
   - Update README [@professor]
   - Adds FFaker::Number.between [@professor]
   - Fixed `FFaker::Number.unique.number(digits: 1)` under Ruby 3 [@simonhildebrandt]